"8:40am. I finally finished the 3 weeks long XCOM campaign on Legendary difficulty yesterday. Then from 10:30pm. I got the urge to read the VR-MCCFR paper. I went to bed past 12am.
As it turns out I can understand the paper now. But some of the things in it are way beyond my expectation. The fact that it does bootstrapping and that bootstraping is unbiased. I just can't believe that last part, not after figuring out what 'bias' means yesterday.

Also, it is a good thing I did not send that mail as I also figured out that CFR+ does strategy averaging as well.

So I do need to figure out strategy averaging. I can't sidestep it.

Does my mind only produce shit?

These algorithms...it is like I cannot see through anything of substance for them at all.

8:50am. Well, let me chill a while and then I will start.

9:30am. Let me finally start.

Yeah, I just do not get it. I told myself to make a language and ML library and succeeded at it, but actually making algorithmic improvements which is the one thing that really matter is something that seems solidly beyond me. Is having insight beyond randomly trying things out impossible?

Just what do I need to do to get it?

https://arxiv.org/abs/1907.09633
Low-Variance and Zero-Variance Baselines for Extensive-Form Games

Let me start by reading this paper. This paper is one of the references on the OpenSpiel page.

Then I will go back to trying to figure out the VR-MCCFR paper's proof of unbiasedness in the appendix.

Along with that, I really need to focus on figuring out the extended to normal form transformation.

"We show that the VR-MCCFR is a specific application of our baseline framework that unnecessarily generalizes across dissimilar states."

9:55am. The predictive baseline is such a brilliant idea.

10:10am. Aaaaaaaaaaaaaaaaaaaaaaahhhhhhhhhhhhh!!!

Why did predictive baselines have to be the optimal thing!!!

Shouldn't conditioning on one sample be an issue!?

Agh...I cannot see anything at all.

And I feel that considering all the histories in the information set goes against the spirit of the thing. Would the predictive baseline be the optimal thing even without that?

"By introducing MCCFR with POS, we provably eliminate the first two sources of variance: the first because we have a zero-variance baseline, and the second because we consider all histories within the information set. For the first time, this allows us to select the MCCFR sampling strategy q t entirely on the basis of minimizing the third source of variance, by choosing the “best” information sets to update. Doing this in a principled way is an exciting avenue for future research. Finally, we close by discussing function approximation. All of the baselines introduced in this paper require an amount of memory that scales with the size of the game tree. In contrast, baseline functions in RL typically use function approximation, requiring a much smaller number of parameters. Additionally, these functions generalize across states, which can allow for learning an accurate baseline function more quickly. The framework that we introduce in this work is completely compatible with function approximation, and combining the two is an area for future research."

This paper is big. For the first time, there is now a viable view for how prediction could work. This baseline is different in character from what came before.

10:20am. It does not feel like I will be getting back into programming any time soon. I want to study the proofs in these two papers in depth.

Yeah, I give up. I can only continue to half ass it until I figure out what I need.

10:25am. "We now define our MCCFR variant, which we call Public Outcome Sampling (POS). Instead of walking trajectories through the EFG tree by sampling actions, POS walks trajectories through the public tree by sampling successor public states."

Wait what?

I thought it was looking into hidden states - like evaluating all the chance nodes in parallel.

17/21. "However, the computations for each information set I ∈ Ii(S) with acting player i can be done completely independently, allowing for easy parallelization (e.g. on a GPU) to achieve complexity O(|I−i(S)||Amax|). This approach was taken with the non-sampling algorithm used in DeepStack [19]."

10:35am. 11/21. This paper has good pseudocode. I can definitely implement this.

I should in fact implement this rather than the more complicated version in the thesis.

10:45am. But just implementing it is not going to give me what I want exactly. I need to do more.

So what is my plan for today?

I need to compose another mail for Lanctot asking him again about that strategy averaging. Apart from that, now that I understand the issues with bootstrapping, I want to spend some time studying the proof in the VR-MCCFR paper.

I really should put that thing into some proof assistant. Agda is too tedious, but Lean has reals and unicode and whatnot.

11:10am. Had to take a little break.

Holy shit, am I asspained right now! Holy shit! I am platinum mad!

First, my attempt to figure out strategy averaging is a failure. Then I figured out what Markovian assumptions really mean in the context of TD learning. Then the VR-MCCFR shows that its bootstrapping is unbiased. Now this paper comes out with predictive baselines.

And the fucking proofs!

Fuck you, just fuck you!

11:15am. Ogh. Is this the fabled bullying? Reality is definitely bullying me here.

This year I've been doing theorem proving for a while, but I fell just a step short of being able to plug proofs like the ones in the paper into a theorem prover. If only I could go a step further in this direction, I could trace through them!

Just looking at those proofs a voice is telling me that I need help for these.

The longing for it is coming full blast at me once again. I dearly wanted to be able to pick some any random math and put it into a proof assitant. Just to see if the rewriting steps make sense.

As a programmer, this should be a basic skill for me.

My theorem proving skills are just a step short of what they need to be! Right now I am just about good enough to be completely useless at what I want to accomplish.

11:20am. Originally I wanted to put PRML through the proof assitant, but talked myself out of it as it would be too much work for too little gain. But these papers are ideal.

11:25am. Ugh, what do I even ask Lanctot at this stage? Let me try composing the email again. I'll have breakfast first though."