"10:15am. epicwisdom: "> Ancient traditions should very well be left behind when they start to get in the way of progress.
This is too general a claim to support your proposals. It seems to me that ML and math are both making great strides in progress, without much new in the way of redefining the foundations of math (yet, at least).

> Personally, I would be happy if proof assistants start pushing out pen and paper. That seems a decent way to start the whole process, and as is usual those who have a need will lead the way.

Who watches the watchers? I don't see proof assistants totally replacing pen and paper until 1) they're ergonomic enough to be as equally expressive as the vernacular notation in every field of mathematics or 2) we get artificial computers implanted in our brains. Certainly this would be a better state of affairs than not using proof assistants, but it is a far off goal, and I don't see any reason to believe it would really say anything surprising about most of the math that is considered "canonical" at present.

> Math has the weakness of being impenetrable, and sometimes even seems like a cultish religion

Or society has the weakness of providing poor education in math. This seems equally ridiculous to me as claiming music theory is "impenetrable." Just because a theory involves obscure knowledge and a steep learning curve does not mean the theory is wrong. It simply means we haven't found a good way to teach it (or more pessimistically, that human brains simply aren't wired to easily learn certain concepts).

> So the moral is - abstract reasoning should not be put on a pedestal, but rather seen as being in service to other areas of cognition. The same goes for math and other fields on the grand scale of things. Going against real world usability is bound to test one's survivability eventually.

You are putting words in my mouth. I am not saying that abstract reasoning is to be put on a pedestal. I am saying that mathematics is a distinct study from other studies, and that pretending otherwise is pointless. What you find mathematical about other studies is merely a borrowing, not a constraint upon mathematics.

Attempting to turn mathematics into physics or computer science is simply a fool's errand; in reality, even if you believe yourself successful, all you will have done is taken a very roundabout path to finding a different mathematical model, without ever touching upon the foundation of mathematics itself. (There are many such"

Note: His post is cut off. I've pasted it correctly.

me: "> It seems to me that ML and math are both making great strides in progress

'we did a bunch of hacks and got 2% improvement on some rigged benchmark' is hardly progress. Running basic RL algorithms on a supercomputer to beat Starcraft/Dota is not progress.

And as for math, isn't it still being taught the same way as it was 100s or even 1000s of years ago? The things learned in high school like calculus is still just 17th century math. And even though we have computers, the main thing that is being graded is manual calculation and following the formulas. Computer algebra systems and proof assistants are a thing now, but schools and universities are still remnants of their middle age selves. Can this really be called progress?

> Who watches the watchers?

I'd much prefer computers to be watching people than people to be watching people. Math as it is done today is just dead letters on a paper. It is an enormous burden to be doing the job of a typechecker and a compiler in one's head. And you need to repeat the effort of learning how to do that for all the fields of mathematics.

The right choice is to offload that burden to the computer. Imagine if you were doing programming the same way you do math - you'd never get anywhere.

> they're ergonomic enough to be as equally expressive as the vernacular notation in every field of mathematics

This is a bit like saying that computers are useless until they reach some arbitrary performance benchmark. Would CPUs be useless until they hit 3Ghz clock speeds? Of course not.

Proof assistants can be judged based on their present quality.

> Or society has the weakness of providing poor education in math.

Forget society. Society at large will always lag at doing what is needed and new initiatives need to be spearheaded. From an objective standpoint, programming itself is actually harder than math, yet there are much more self-taught great programmers than there are mathematicians.

The reason for why that is because putting the computer in the learning loop provides feedback and turns the process into a game. That is what proof assistants do and need to continue doing when it comes to math.

> It simply means we haven't found a good way to teach it (or more pessimistically, that human brains simply aren't wired to easily learn certain concepts).

Or maybe PA/PLs simply weren't good enough until now.

If you honestly care about people learning math rather than just memeing about it, you would want to make it so that math is as useful as it could be in the real world. Real world use translates directly to study motivation. Students working to parrot what the teacher wants to hear is the opposite of that. Not many with actual intelligence going through the school system would put their faith in it.

Faith is important and computers are more trustworthy than people. And computability itself is directly related to real world use. In fact, the relationship is much stronger than that - without computability, you cannot have real world use. People can sense that deep down.

One argument that I often hear is that progress cannot be anticipated and that moonshot research is necessary, especially in pure math. There are some fields like number theory which were thought by their founders that they will never have a practical use and yet turned out to have.

But number theory specifically always had a computable character. It is not one of the 'troublesome' fields to borrow Wildberger's words. If it were, you could not compute and do cryptography as a result.

Quite a lot of the fields of pure mathematics do not have a computable character. The difficulties of computation are buttressed and swept under the carpet and for what?

When that happens, math turns from being useful to a bunch of people sharing stories to each other. The game just becomes about posturing and appearing smart. It is a prime example of corruption of peer review.

This is more than a dichotomy between pure and applied math. Since pure math is the more abstract part and harder to ground in real world use, it is actually the one which needs to be more grounded in logic instead.

> I am saying that mathematics is a distinct study from other studies, and that pretending otherwise is pointless.

This is a bug rather than a feature. It should be fixed.

If needed, math itself should be devoured by computer science and have its essence cleansed and converted into something more useful."

10:20am. I am enjoying this too much. Mrkkk...I guess this is a decent way to start the morning.

10:25am. I really should get started on the manual. Let me do a bit brief of reading and then I will do so."