"10:05am. I am back. So yesterday I finally finished MGRP Restart. The later half was pretty intense. I wish the thing would get animated, the world needs more magical girl death match anime. And right now I am downloading Pandora in the Crimson Shell vol 11.
The last two days were a rare break for me. I did do a bit of work - in fact in some way doing only a few hours of steady work per day is ideal, but now it is time to get back into the swing of things. I want to finish CFR. And it is only when I finish it that the real deal will start.

I made Spiral, but only in a dependently typed PL like Lean can I implement not just the algorithms, but the theorems themselves.

10:10am. It will be the first time I am trying to do something like that.

Dead letter theorems in some paper are meaningless. It is only that when they are woven into the fabric of the program itself that you can consider the task being completed.

10:10am. I do not think anybody else in the world tried something like this up to now, so I will be making new strides in this.

At the current time, there was some work on using ML to prove theorems, but nothing like actual algorithms implemented and proven in some prover.

This whole thing will take a while and I do not know if it will get me anywhere, but trying things out at random like the field at large is doing is just too unprofitable for me. The only chance I have is to sharpen my reasoning and find a way to cut through. Unless I do that I won't be able to meet the bare minimum needed for ignition of the Singularity.

To go down this route, I really am desperate. I am so alone and far away from home here.

Yet, I must continue fighting.

10:15am. Ghost Urn and Limited can wait for later. Let me see if I got any replies to the questions I posted and then I will work from there."