"Let me watch that vid.
6:50pm. https://www.youtube.com/watch?v=VUdFdlQNfpg&feature=youtu.be

It was actually quite funny. Let me watch the next one.

7:05pm. https://youtu.be/Q3V9UNN4XLE?t=578
"The beginning of any subject of mathematics is the most important time for that subject."

Yes. This is what I've been looking for some body to day.

This is the first time I've ever seen somebody say that the foundation is what really matters. If I apply this philosophy to ML, I will be able to win.

7:25pm. https://youtu.be/Q3V9UNN4XLE?t=991

Here he talks about the future of mathematics. The way he talks about number theory actually makes it sound like a subject I might be interested in.

"Higher dimensional geometry is going to be a huge subject of which we know almost nothing about. We are vastly ignorant of higher dimensional geometry really. And there is so much work to be done in this direction. Just astounding."

"There is going to be a much bigger role for number theory. Number theory has been for a long time been a rather exotic subject that has been pursued for its own internal beauty. Lately now, it has been more connected to things having to do with coding and encryption and so on. But I think we are going to see number theory playing a role, much more of a role when it becomes clear that approximate statements that the real number orientation gives us towards the solutions to equations, intersections of curves of varieties. That these are really replaced with number theoretical issues. When you go from floating point to full integer precision then the solution of an equation is often a number theoretical issue. Whereas before it was a approximation issue.

Number theory is going to be one of those subjects which naturally expands with this rational point of view."

This is completely new to me. This is something I am definitely going to have to look into.

7:55pm. Let me finally relax a little. Sheesh."