"12:55pm. https://www.reddit.com/r/MachineLearning/comments/donbz7/d_im_so_sick_of_the_hype/
Interesting thread here.

1:05pm. Let me watch the latest ep of Average and then I will resume.

1:40pm. Let me resume.

2:10pm. "Hello again.
I've been busy getting familiar with the Lean prover, and am ready to start trying to formalize some of the proofs in the CFR papers seriously. I've opened a repo to track my progress here.
Right now my focus is on doing the 'Regret Minimization in Games with Incomplete Information' paper. I've translated the first theorem from the appendix here. This can only be considered a sketch at this point and the proof has many holes that need filling. In additions, the assumptions I've made axiomatically need to be given constructively for the proof to be rigorous. Here is how far I've gotten with that.
As can be expected, I am running into difficulties given the lack of precise definition for the many of the things in the paper.
If the task was simply to implement an algorithm, I could go from the pseudo-code and the actual code and sidestep the difficulties with the definitions being loose, but formal proofs have stricter requirements. As things are now I cannot reproduce even the first step of the lemma 6 which the author gave as-is presuming it to be obvious.
Let me list out some of the pain points.
1) The way to implement the utility calculation is ambiguous. If you think about it, the player, the policy and the infoset is not enough information to actually calculate the utility, you also need the game tree as well. This by itself is not to painful - you can just imagine the utility being calculated for a particular game tree.
The ambiguous part here though is that the game tree might have multiple nodes aliased to the same infoset and each of them might have their own different utility. Presumably those utilities should be summed up. But how should they be weighted?
In the actual CFR algorithm, they are weighted by the opponent's reach probability, but the algorithm itself is unambiguous in the way those calculations should be done. That is because the notion of traversing over a game tree is explicit, so the utility calculation is for the node, given the other parameters.
I can't see a way to make the proof work given the way utility is calculated in the actual algorithm.
2) The way reach probability is to be calculated in the proof is ambiguous as well. The proof structure does not indicate that the underlying functions are traversing a tree. Sure, it would be easy to sum up all the reach probabilities given an infoset over the game tree, but that would not be right.
3) Let me bring up eq 8 again as this is the biggest troublemaker. You said the equality should be an inequality, but regardless of that, the proof that this particular step holds is not the provided and it is not trivial to fill in. In fact, I cannot think of a single way to do it.
---
There is more, but I might be able to make some progress if I could precisely define what the various functions are.
Sorry for bothering you about this given that this is not your paper and you weren't the one who made the proofs. If you'd prefer I could mail some the authors from the paper specifically. I did mail a few questions to Zinkevich a month ago, but did not receive a reply. I guess he might not be interested in talking about his academic work given that the last few entries (from 2013) talk of him being at Yahoo. Given that the last update was from that time, I hope this is a sign of stable employment and not death.
Also, I haven't yet had the chance to look into any of the strategy averaging related links from last time as I was busy with getting familiar with Lean. I'll be sure to do it after I am done with the theorem proving.
- Marko Grdinić"

This should do it. There should also be links here, but copy paste did not catch them.

I'll send this. Let me stop here as I need to do a chore. I'll get back into the swing of things soon. While I wait for a reply, I'll read some different papers or maybe watch some vids by Wildberger."