"1:35pm. I feel exhausted somehow despite not doing any programming. This is as a result of yesterday and today.
I am a bit stressed out right now over moving from Lean to Agda. Which is right?

What do I want to accomplish? What would be better for me, to implement the proofs in Lean or Agda?

I said I would switch due to the standard library issues...but is it really such a big problem to implement the Vec functions on my own? It is not that this would be a big deal.

Forgoing the libraries, which language would be better?

2pm. https://lean-forward.github.io/logical-verification/2018/41_notes.html
"What is the difference between Prop and Type?

Each proposition (when interpreted as a type) is either empty or has exactly one inhabitant: Prop = {∅, {⋆}}

∀(p : Prop) (h₁ h₂ : p), h₁ = h₂
This is called proof irrelevance and makes reasoning about dependent types easier

Proof irrelevance is implemented in the kernel, which gives us definitional equality, allowing proofs by refl

In contrast:

Agda and Coq are proof relevant by default (but are compatible with proof irrelevance)

Homotopy type theory and other constructive or intuitionistic type theories build on data in equality proofs and therefore are incompatible with proof irrelevance

Prop is an instance of subsingleton α := ∀a b : α, a = b

(Other subsingletons: unit, empty, decidable P, ...)"

2:15pm. https://doisinkidney.com/posts/2018-09-20-agda-tips.html
"Irrelevance

This one is a little confusing, because Agda’s notion of “irrelevance” is different from Idris’, or Haskell’s. In all three languages, irrelevance is used for performance: it means that a value doesn’t need to be around at runtime, so the compiler can elide it.

That’s where the similarities stop though. In Haskell, all types are irrelevant: they’re figments of the typechecker’s imagination. You can’t get a type at runtime full stop.

In dependently typed languages, this isn’t a distinction we can rely on. The line between runtime entities and compile-time entities is drawn elsewhere, so quite often types need to exist at runtime. As you might guess, though, they don’t always need to. The length of a length-indexed vector, for instance, is completely determined by the structure of the vector: why would you bother storing all of that information at runtime? This is what Idris recognizes, and what it tries to remedy: it analyses code for these kinds of opportunities for elision, and does so when it can. Kind of like Haskell’s fusion, though, it’s an invisible optimization, and there’s no way to make Idris throw a type error when it can’t elide something you want it to elide.

Agda is totally different. Something is irrelevant in Agda if it’s unique. Or, rather, it’s irrelevant if all you rely on is its existence. It’s used for proofs that you carry around with you: in a rational number type, you might use it to say that the numerator and denominator have no common factors. The only information you want from this proof is whether it holds or not, so it’s the perfect candidate for irrelevance.

Weirdly, this means it’s useless for the length-indexed vector kind of stuff mentioned above. In fact, it doe exactly the opposite of what you might expect: if the length parameter is marked as irrelevant, the the types Vec A n and Vec A (suc n) are the same!

The way you can use it is to pattern-match if it’s impossible. Again, it’s designed for eliding proofs that you may carry with you otherwise."

2:25pm. https://agda.readthedocs.io/en/v2.6.0.1/language/irrelevance.html

Ah, so that is what that . in the rational means.

2:45pm. https://hal.inria.fr/hal-01859964/
Definitional Proof-Irrelevance without K

I am just mentally so overburdened right now. Let me do the chores and I will read this paper. Agda's Prop are based on it.

2:50pm. I was wrong. Looking at the examples in the Agda documentation, I realize now that proof-irrelevance has a very real purpose.

Let me do the chores rather than just sitting here thinking."