"3:20pm. https://leanprover.github.io/theorem_proving_in_lean/axioms_and_computation.html#quotients https://leanprover.github.io/talks/Lean_BPR.pdf
Let me read these slides.

As I said, right now I am overburdened. I am trying to think through my ordeal logically, I can't really justify switching to Agda on the basis of the Prop issues and the standard library.

The Prop issues which I attributed to Lean, feel more like my own fault right now. In fact that is what they are. I sort of assumed that Lean should be something, while in fact it was something else.

3:30pm. "Definitional equality—or conversion—for a type theory with a decidable type checking is the simplest tool to prove that two objects are the same, letting the system decide just using computation. Therefore, the more things are equal by conversion, the simpler it is to use a language based on type theory. Proof-irrelevance, stating that any two proofs of the same proposition are equal, is a possible way to extend conversion to make a type theory more powerful. However, this new power comes at a price if we integrate it naively, either by making type checking undecidable or by realizing new axioms—such as uniqueness of identity proofs (UIP)—that are incompatible with other extensions, such as univalence. In this paper, taking inspiration from homotopy type theory, we propose a general way to extend a type theory with definitional proof irrelevance, in a way that keeps type checking decidable and is compatible with univalence. We provide a new criterion to decide whether a proposition can be eliminated over a type (correcting and improving the so-called singleton elimination of Coq) by using techniques coming from recent development on dependent pattern matching without UIP. We show the generality of our approach by providing implementations for both Coq and Agda, both of which are planned to be integrated in future versions of those proof assistants."

So Coq is going to get this as well.

3:40pm. "An alternative approach is to do as in Lean, where they do have proof irrelevance with singleton elimination, but they only implement a partial version of proof irrelevance for recursive inductive types satisfying the singleton elimination, which is restricted to closed terms.2 But this partial implementation of the conversion algorithm breaks in particular subject reduction, which seems a desirable property for a proof assistant (see a concrete example in Appendix A)."

4:05pm. Ok, I am done with this. Basically, I am going to go with Lean, at least for CFR.

I've reasoned it out.

Basically if I want to do equality proofs, then tactics are what I need, period. Furthermore Lean's `calc` blocks, are simply needed here.

I feel that programming can be split into two parts - actual computations and proofs of these equalities. For the later I need tactics unless I want to go through a lot of annoyance.

So from the perspective of just that activity, Lean is better than Agda.

I had pains with typeclasses as well, but like automation, typeclasses are not an disadvantage. I'll assume that Lean and Agda are evenly matched here now that Agda has those new instance arguments.

Then comes the standard library.

Yes, I am definitely pissed at Lean here for not having a proper `Vec` amongst other things. And for forcing me to implement those custom induction principles for fin.

Still that is about it. It is not like it is difficult for me to implement my own custom functions. There are just lists we are talking about here. I just need to make up my mind and they will happen.

4:15pm. Then `Prop`. As I noted, there is something wrong considering it to be a disadvantage. I will accept that the difficulties I had with it, are due to my own lack of skill.

4:20pm. Still, that does not remove my objections that the handling of `Prop`'s is awkward in Lean. I am going to change some of my complaints given the new information by Mario Carniero, but I'll still leave it in Bad.

Ok, here is how I need to decide this.

I am going to assume that Agda and Lean are roughly equal in terms of type system features. I'll assume that typeclasses vs instance arguments are roughly even. I am going to ignore Lean being somewhat awkward with the way it compiles to recursors. I am assuming that is where the Prop issues stem from.

I am going to assume that standard libraries are kind of in the same shape. Right now Agda is giving trouble with that <= operator, while Lean is doing everything wrong with the types I want to use.

4:25pm. I will remember what I once read, and assume that whatever differences are there between proof assistants, that I can just finesse over them.

I will ignore the syntax differences.

4:30pm. So the final decision really comes down to one thing - in the absence of big type system differences, like cubical type theory that Cubical Agda has.

The final decision really comes down to - do I want or do I not want tactics?

In other words, do I want a language for implementing things, or do I want a language for proving things?

...Here I am trying to prove things so I am going to go with Lean.

Admittedly, Agda is not much of a language for implementing things at the moment - none of the DT langs are, but future dependently typed languages will be. Maybe I'll give Spiral dependent types at some point. Hopefully that point will be past the self improvement loop.

6:15pm. Done with lunch. There are days like today when I get fully obsessed over design issues.

But this whole thing is finally forcing me to state what do I really think about dependent types after six months of experience with them. Since I have a fork in the road, it is pushing to me to take a view. What is right? And what should the criteria for being right be.

6:25pm. I am starting to have some ideas about what the right way to look at dependent types should be.

I'll call it a day here. I need to turn off somehow and allow myself to unwind."