"6:50pm. I guess when in doubt, go with beauty.
6:55pm. That `Vec.foldl` broke me today. Somehow I am really grateful that the monthly thread is delayed.

7:25pm. https://mangadex.org/title/32386/mato-seihei-no-slave

A new chapter of this is out and I am too depressed to get into it.

7:30pm. https://antonhowes.substack.com/p/age-of-invention-rise-of-the-mathematicians

https://antonhowes.substack.com/p/age-of-invention-the-crucial-century?utm_source=substack&utm_content=topposts

This is pretty interesting.

"If a peaceful extraterrestrial visited the world in 1550, where would it predict would be the most likely site for an Industrial Revolution? That’s the subject of my latest blogpost, “The Crucial Century”, which you can read here. I run through a number of potentially important factors, including urbanisation, major cities, living standards, military might, and state capacity. But on all of them, England was unexceptional, really even a backwater."

8:05pm. I think I am going to go for Agda after all. Today's botched attempt at trying to implement `foldl` is proof enough that Lean is toying with me.

As pitiful as it is to cut an effort after investing 3 weeks in it, I'll do it.

8:10pm. Let me add a bit to the review.

9:20pm.

///

Agda vs Lean:

I am really torn between Agda and Lean in which at this point I roughly have equal amount of experience. Agda has beauty on its side which a strength of its own. In the world of PL, beauty definitely has a connection with power. And in the worst case, when used as a decision heuristic beauty won't steer you wrong by too much.

On paper, Lean's design should be superior. Automation on its own cannot be a disadvantage. But there is a big gap between being in the loop and having built that automation yourself, and somebody coming in from the outside. Lean's arrows all point in the right direction, but it will take some significant work to tighten up the design where an entrant is not stumbling into surprise after surprise. Lean is ambitious. It feels like it is taking on large commitments and it remains to be seen whether it will be able to meet them.

Agda on the other hand has a narrower focus, seemingly purposely avoiding the kind of utility that having tactics would provide. The end result is that the thing it does do well are better than Lean. For example, the overall structure of the library code is cleaner in Agda. Its pattern matching is definitely better. And its design feels more coherent. What that means is that common patterns of use are optimized to be as ergonomic as possible. Agda is definitely easier for the student, I can vouch for that based on personal experience.

Which is better? Despite knowing exactly what I want, I can only apologize that at this point I have no clear answer. I cannot sort the conflicting thoughts floating through my mind. I can't weight the potential utility of tactics against ergonomics at my level.

///

I added the above to the review. I guess I'll make it my third post. Yeah, since I am confused, I'll just go with the beauty heuristic.

Tomorrow, it is Agda time. I am really greatly stressed here, so let me call it a day.

11/2/2019

8:25am. I am up. Let me chill a little. I do not feel like starting right away. In fact, let me have breakfast now.

9:35am. Let me finally start. I already updated the SO post. Let me link to it in chat and then I am done with. After wasting two days on thinking whether I should switch languages or not, I do not care anymore. What is important is that I sit down and start working on those damn proofs.

9:50am. https://stackoverflow.com/questions/9472488/differences-between-agda-and-idris?rq=1

Right now I am reading some SO questions. But let me start. If I need heterogenous equality I will bring it out.

9:55am.

```
------------------------------------------------------------------------
-- The Agda standard library
--
-- Automatic solvers for equations over naturals
------------------------------------------------------------------------

-- See README.Nat for examples of how to use this solver

{-# OPTIONS --without-K --safe #-}

module Data.Nat.Solver where

import Algebra.Solver.Ring.Simple as Solver
import Algebra.Solver.Ring.AlmostCommutativeRing as ACR
open import Data.Nat using (_≟_)
open import Data.Nat.Properties

------------------------------------------------------------------------
-- A module for automatically solving propositional equivalences
-- containing _+_ and _*_

module +-*-Solver =
  Solver (ACR.fromCommutativeSemiring *-+-commutativeSemiring) _≟_
```

Oh Agda does have solvers. I guess it is not true that it does not have them.

10am. https://oisdk.github.io/agda-ring-solver/README.html

Actually, these solvers are pretty much the main reason why I was considering Lean over Agda. It is just so tedious to do all those operations by hand. But Agda can do this bit of drudgery for me, then I have much less reason to avoid it.

Let me read the paper.

10:10am.

```
  lemma₁ : ∀ x y → x + y * 1 + 3 ≈ 3 + 1 + y + x + - 1                        --   ████║ ██║  ██║
  lemma₁ = solve Int.ring                                                     -- ██████████║  ██║
                                                                              --   ████╔═██║  ██║
  lemma₂ : ∀ x y → (x + y) ^ 2 ≈ x ^ 2 + 2 * x * y + y ^ 2                    --     ██║ ██║  ██║
  lemma₂ = solve Int.ring
```

This is fantastic. I had no idea Agda would be capable of this. Maybe there is hope for it after all. It can even generate step by step proofs.

10:15am. `* Did you know that Agda has [ring solvers](https://oisdk.github.io/agda-ring-solver/README.html)? I sure didn't till now!`

I'll remove the review I did yesterday. It seems until I get enough exp on both, any model I have of the languages will necessarily be off. But this does move things to the side of Agda being the right choice for me. It makes it a lot more appealing.

Ok. So I might have spent essentially most of the last month playing around, but I did learn a bunch of stuff.

Let me see if I can get serious in November."