"10:35am. Oh, it seems Agda has inbuilt Floats! I just found it in the standard library. What about fixnums?
It has built in Nats. I wonder if they are optimized. I think that they are in Idris.

https://agda.readthedocs.io/en/v2.6.0.1/language/built-ins.html

```
The effect of binding the boolean type is that you can then use primitive functions returning booleans, such as built-in NATEQUALS, and letting the GHC backend know to compile the type to Haskell Bool.
```

```
* The use of natural number literals is enabled. By default the type of a natural number literal will be Nat, but it can be overloaded to include other types as well.
* Closed natural numbers are represented as Haskell integers at compile-time.
* The compiler backends compile natural numbers to the appropriate number type in the target language.
* Enabled binding the built-in natural number functions described below.
```

What are Haskell integers?

https://hackage.haskell.org/package/integer-gmp-0.5.1.0/docs/GHC-Integer.html

It seems they are arbitrary precision.

10:50am. Let me finally start work on that datatype.

How did I imagine it?

```
data UniqueList {a} (A : Set a) : Set a
data UniqueElem {a} (A : Set a) : A → UniqueList A → Set a
```

Now let me define this.

```
data UniqueList {a} (A : Set a) : Set a
data UniqueElem {a} {A : Set a} : A → UniqueList A → Set a

data UniqueList A where
  [] : UniqueList A
  _∷_ : ∀ (x : A) (xs : UniqueList A) → UniqueElem x xs → UniqueList A
```

Wonderful. It is really coming together.

```
data UniqueList A where
  [] : UniqueList A
  _∷_ : ∀ (x : A) (xs : UniqueList A) → ¬ UniqueElem x xs → UniqueList A

data UniqueElem {a} {A} where
  here : ∀ (x : A) (xs : UniqueList A) → UniqueElem x (x ∷ xs)
```

Now wait, shit it is not. It completely slipped my mind that now this has 3 constructors.

11:10am.

```
data UniqueList A where
  [] : UniqueList A
  cons : ∀ (x : A) (xs : UniqueList A) → ¬ UniqueElem x xs → UniqueList A

data UniqueElem {a} {A} where
  here : ∀ (x : A) (xs : UniqueList A) prf → UniqueElem x (cons x xs prf)
```

```
E:\TheoremProving\Agda stuff\hello.agda:3,6-11,74
UniqueList is not strictly positive, because it occurs
in the type of the constructor here
in the definition of UniqueElem, which occurs
to the left of an arrow
in the type of the constructor cons
in the definition of UniqueList.

E:\TheoremProving\Agda stuff\hello.agda:3,6-11,74
UniqueElem is not strictly positive, because it occurs
to the left of an arrow
in the type of the constructor here
in the definition of UniqueElem.
```

This is not working. I am doing something wrong.

```
data UniqueList A where
  [] : UniqueList A
  cons : ∀ (x : A) (xs : UniqueList A) → UniqueElem x xs → UniqueList A

data UniqueElem {a} {A} where
  here : ∀ (x : A) → UniqueElem x []
```

Let me try doing it directly.

11:25am.

```
open import Relation.Nullary
open import Relation.Binary.PropositionalEquality

data UniqueList {a} (A : Set a) : Set a
data NotIn {a} {A : Set a} : A → UniqueList A → Set a

data UniqueList A where
  [] : UniqueList A
  _∷_ : ∀ (x : A) (xs : UniqueList A) → NotIn x xs → UniqueList A

data NotIn {a} {A} where
  done : ∀ (x : A) → NotIn x []
  dive : ∀ (x y : A) (ys : UniqueList A) prf → NotIn x ys → ¬ x ≡ y → NotIn x ((y ∷ ys) prf)
```

This is it. Definitely. It is what I've been looking for.

11:30am.

```
data NotIn {a} {A} where
  done : ∀ (x : A) → NotIn x []
  dive : ∀ (x y : A) (ys : UniqueList A) prf → NotIn x ys → ¬ x ≡ y → NotIn x ((y ∷ ys) prf)
```

No wait. There is something wrong here. What is prf supposed to be here exactly?

```
data NotIn {a} {A} where
  done : ∀ (x : A) → NotIn x []
  dive : ∀ (x y : A) (ys : UniqueList A) (prf : NotIn y ys) → NotIn x ys → ¬ x ≡ y → NotIn x ((y ∷ ys) prf)
```

Ah, it is this thing.

11:40am.

```
data UniqueList A where
  [] : UniqueList A
  _∷_ : ∀ (x : A) (xs : UniqueList A) → NotIn x xs → UniqueList A

data NotIn {a} {A} where
  done : ∀ (x : A) → NotIn x []
  dive : ∀ (x y : A) (ys : UniqueList A) (prf : NotIn y ys) → NotIn x ys → ¬ x ≡ y → NotIn x ((y ∷ ys) prf)
```

I really am completely confused here. It feels way more complex than it should. Let me try the normal definition.

11:50am.

```
data _[_]=_ {A : Set a} : ∀ {n} → Vec A n → Fin n → A → Set a where
  here  : ∀ {n}     {x}   {xs : Vec A n} → x ∷ xs [ zero ]= x
  there : ∀ {n} {i} {x y} {xs : Vec A n}
          (xs[i]=x : xs [ i ]= x) → y ∷ xs [ suc i ]= x
```

Isn't something like this defined in the standard library for Agda? I can't find it.

12:05pm. https://stackoverflow.com/questions/58679537/what-would-be-a-simple-implementation-of-a-collection-with-non-repeating-element
https://stackoverflow.com/questions/58679462/is-there-an-element-in-list-datatype-defined-in-the-standard-library

12:10pm. This is so troublesome. I need to think about this more. I thought I could just sidestep trees, but trying to better than my Coq effort is proving hard.

2:05pm. Let me resume. I haven't really rested during the break.

What I am going to do now is simple. Go through all the List files for that elemental membership definition. I just can't believe that the definition is not in the library. I just need to look for it harder. Then I will implement the uniqueness of elements in the most straightforward way possible.

```
tabulate : ∀ {n} (f : Fin n → A) → List A
tabulate {n = zero}  f = []
tabulate {n = suc n} f = f zero ∷ tabulate (f ∘ suc)
```

Oh this is an interesting way of implementing this. Yeah, this is how I could have done the lean thing.

```
data Any {A : Set a} (P : Pred A p) : Pred (List A) (a ⊔ p) where
  here  : ∀ {x xs} (px  : P x)      → Any P (x ∷ xs)
  there : ∀ {x xs} (pxs : Any P xs) → Any P (x ∷ xs)
```

Ah, I think this is what I am looking for.

2:30pm.

```
data NonRepeating {a} {A : Set a} : (l : List A) → Set a where
  done : NonRepeating []
  rest : ∀ {x xs} → ¬ Any (x ≡_) xs → NonRepeating xs → NonRepeating (x ∷ xs)

record UniqueList {a} (A : Set a) : Set a where
  constructor _//_
  field
    l : List A
    wf : NonRepeating l
```

I think I will go with this. Let me update the SO answers.

2:40pm.

```
data NonRepeating {a} {A : Set a} : (l : List A) → Set a where
  done : NonRepeating []
  rest : ∀ {x xs} → ¬ Any (x ≡_) xs → NonRepeating xs → NonRepeating (x ∷ xs)

record UniqueList {a} (A : Set a) : Set a where
  constructor _//_
  field
    l : List A
    wf : NonRepeating l

qwe : ∀ {a} {A : Set a} → UniqueList A → Set
qwe (.[] // done) = {!!}
qwe (.(_ ∷ _) // rest x wf) = {!!}
```

This works great. As I said, let me update answer.

2:45pm. Done with the updates. Let me give it a try.

I'll make the UniqueList my set of infosets. I'll key the tree to them. I'll do this from the top.

2:50pm. I guess this is the power of hierarchies. Something like the above was in fact my first idea, but I could not come up with anything better than it. It is a remarkable simplification compared to the mutually recursive formulation.

```
data NonRepeating {a} {A : Set a} : (l : List A) → Set a where
  done : NonRepeating []
  rest : ∀ {x xs} → ¬ Any (x ≡_) xs → NonRepeating xs → NonRepeating (x ∷ xs)

record UniqueList {a} (A : Set a) : Set a where
  constructor _//_
  field
    l : List A
    wf : NonRepeating l
```

But even more than this, I need an associative list.

3pm. https://stackoverflow.com/questions/42614042/map-with-strings-as-keys-in-agda

Hmmmhhh...

Should I put some effort in getting familiar with AVL trees?

https://www.youtube.com/watch?v=FNeL18KsWPc
6. AVL Trees, AVL Sort

Ah, what the hell. Let me spend some time getting familiar with this. If it does not work, I'll do my own associative list."